Add aoc_mod.
aoc_mod is a module with Python functions for advent of code. It can be imported in notebooks, and abstracts some code I repeat a lot.

Update the README for day2 instructions. Update the main README with web UI instructions.

Update the run jupyter script, add building and installing for aoc_mod